update/improve_ui
