+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:theme="@style/Theme.BasicRecyclerViewExample">
 
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true"
             android:label="@string/str_my_fav_movies">
             <intent-filter>
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MovieDetailActivity"
+            android:name=".activity.MovieDetailActivity"
             android:exported="true"
             android:label="@string/str_movie_detail" />
     </application>
